djangobackend/requirements.txt
asgiref==3.5.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
